+++ chronosphere/tfschema/slo.go
@@ -102,6 +102,21 @@ var SloDefinition = map[string]*schema.Schema{
 			},
 		},
 	},
+	"time_window": {
+		Type:     schema.TypeList,
+		Optional: true,
+		Computed: true,
+		MinItems: 1,
+		MaxItems: 1,
+		Elem: &schema.Resource{
+			Schema: map[string]*schema.Schema{
+				"duration": {
+					Type:     schema.TypeString,
+					Required: true,
+				},
+			},
+		},
+	},
 	"burn_rate_alerting_config": {
 		Type:     schema.TypeList,
 		Computed: true,
@@ -109,6 +124,11 @@ var SloDefinition = map[string]*schema.Schema{
 		Elem: &schema.Resource{
 			Schema: BurnRateDefinition,
 		},
+	},
+	"enable_burn_rate_alerting": {
+		Type:     schema.TypeBool,
+		Optional: true,
+		Computed: true, // This has to be computed as it's being backfilled based on other flags.
 	},
 }