+++ examples/slo/main.tf
@@ -33,6 +33,7 @@ resource "chronosphere_slo" "slo" {
     reporting_windows {
       duration = "28d"
     }
+    enable_burn_rate_alerting = true
   }
 
   sli {
@@ -79,6 +80,7 @@ resource "chronosphere_slo" "slo_with_signal_grouping_signal_per_series" {
       budget = 99
       severity = "warn"
     }
+    enable_burn_rate_alerting = true
   }
 
   sli {
@@ -103,6 +105,7 @@ resource "chronosphere_slo" "slo_with_signal_grouping_labels" {
     reporting_windows {
       duration = "28d"
     }
+    enable_burn_rate_alerting = true
   }
 
   sli {
@@ -123,6 +126,10 @@ resource "chronosphere_slo" "slo_with_filters" {
     reporting_windows {
       duration = "28d"
     }
+    time_window {
+      duration = "28d"
+    }
+    enable_burn_rate_alerting = true
   }
 
   sli {
@@ -154,3 +161,28 @@ resource "chronosphere_slo" "slo_with_filters" {
     }
   }
 }
+
+resource "chronosphere_slo" "slo_without_alerting" {
+  name                   = "SLO Without Alerting"
+  collection_id          = chronosphere_collection.c.id
+  notification_policy_id = chronosphere_notification_policy.np.id
+
+  definition {
+    objective = 99.95
+    reporting_windows {
+      duration = "28d"
+    }
+    time_window {
+      duration = "28d"
+    }
+  }
+
+  sli {
+    custom_indicator {
+      bad_query_template   = "sum(rate(http_request_duration_seconds_count{error=\"true\"}[{{ .Window }}]))"
+      total_query_template = "sum(rate(http_request_duration_seconds_count[{{ .Window }}]))"
+    }
+    custom_dimension_labels = ["label1", "label2"]
+  }
+}
+